requirements.txt
asgiref==3.5.2
Django==4.1.2
django-restframework==0.0.1
djangorestframework==3.14.0
pytz==2022.5
sqlparse==0.4.3
